fix: OBO error when calculating global scope
When we added nloc on top of lizard (as a function) and then
calculated global scope in PyDiffer, we got an Off-By-One
error due to that we added global scope in lizard,
and then subtracting that global scope from the global scope in
pydiffer. Solution is to remove global scope from lizard
since its not used.

Signed-off-by: Jesper Derehag <jderehag@hotmail.com>