Fix: Querying emails cleared user_id
When updating emails, it would clear user_id column of both
defects and change_metrics. This was due to that sqlaclchemy used
cached copies of Users. So when deleting said user, it would
also clear the defect & change_metric columns.
So instead defer deletion, and make sure to flush any caches
apriori.

Signed-off-by: Jesper Derehag <jderehag@hotmail.com>