Increased size of VARBINARY
Some functions are apparantly larger than 255 chars. In this case it meant that
they where properly stored in sqlite (>255 due to that sqlite ignores that boundary).
But when mirroring to mysql it meant that those functions got truncated and got duplication error.

Easiest solution is simply to increase size of VARBINARY.

Signed-off-by: Jesper Derehag <jderehag@hotmail.com>