Add support for quick consistency checking
Add functionality that checks that all files are covered by
exactly one rule in the maintainers file. A new method, verify(),
is added to Maintainers, which will return a list of all files
that do not match this criteria.

In order to do this quickly, the Maintainers class will contain
a tree structure of all the rules in a maintainer file. This will
avoid having to check the entire list of rules for every file.