Fix: Git: Infinite recursion when looking up symbolic links
In the file cache we would end up in an infinite recursion if the file
to look up was a symbolic link. First we check if its in the cahce,
failing that we will update the file_cache (but now for a normalized
path, pointing towards the real file), but we will still look up the
symbolic file, ending up in an infinite recursion.

Signed-off-by: Jesper Derehag <jderehag@hotmail.com>